requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
asgiref==3.5.2
    # via django
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via edx-django-utils
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   drf-yasg
    #   openapi-codec
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==37.0.4
    # via
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.14
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
django-cors-headers==3.13.0
    # via -r requirements/base.in
django-crum==0.7.9
    # via edx-django-utils
django-extensions==3.2.0
    # via -r requirements/base.in
django-model-utils==4.2.0
    # via -r requirements/base.in
django-rest-swagger==2.2.0
    # via -r requirements/base.in
django-waffle==2.6.0
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.13.1
    # via
    #   -r requirements/base.in
    #   django-rest-swagger
    #   drf-jwt
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via edx-drf-extensions
drf-yasg==1.21.3
    # via edx-api-doc-tools
edx-api-doc-tools==1.6.0
    # via -r requirements/base.in
edx-auth-backends==4.1.0
    # via -r requirements/base.in
edx-django-release-util==1.2.0
    # via -r requirements/base.in
edx-django-utils==5.0.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.0.1
    # via -r requirements/base.in
edx-opaque-keys==2.3.0
    # via edx-drf-extensions
edx-rest-api-client==5.5.0
    # via -r requirements/base.in
future==0.18.2
    # via pyjwkest
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
mysqlclient==2.1.1
    # via -r requirements/base.in
newrelic==7.16.0.178
    # via edx-django-utils
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via django-rest-swagger
packaging==21.3
    # via drf-yasg
pbr==5.9.0
    # via stevedore
psutil==5.9.1
    # via edx-django-utils
pycparser==2.21
    # via cffi
pycryptodomex==3.15.0
    # via pyjwkest
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==2.4.0
    # via
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pymongo==3.12.3
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via edx-drf-extensions
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.1
    # via
    #   -r requirements/base.in
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via edx-django-release-util
requests==2.28.1
    # via
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
semantic-version==2.10.0
    # via edx-drf-extensions
simplejson==3.17.6
    # via django-rest-swagger
six==1.16.0
    # via
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   pyjwkest
    #   python-dateutil
slumber==0.7.1
    # via edx-rest-api-client
social-auth-app-django==5.0.0
    # via edx-auth-backends
social-auth-core==4.3.0
    # via
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.2
    # via django
stevedore==4.0.0
    # via
    #   edx-django-utils
    #   edx-opaque-keys
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.10
    # via requests
